# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/entity/children": {
            "get": {
                "description": "Get all children of a specific entity, with optional recursion and filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Entity Management"
                ],
                "summary": "Get entity children",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cognito ID",
                        "name": "X-Cognito-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Whether to include all descendants",
                        "name": "recursive",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum depth level for descendants (0 for direct children only, -1 for all)",
                        "name": "level",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by category type",
                        "name": "category_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entity children retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.EntityChildrenResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Entity not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/entity/hierarchy": {
            "get": {
                "description": "Get an entity and all its descendants as a hierarchical structure",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Entity Management"
                ],
                "summary": "Get entity hierarchy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cognito ID",
                        "name": "X-Cognito-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum depth to include (default: 10)",
                        "name": "max_depth",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entity hierarchy retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.EntityHierarchyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Entity not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/entity/root": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new top-level entity with optional user association",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Entity Management"
                ],
                "summary": "Create a root entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cognito ID",
                        "name": "X-Cognito-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Entity information",
                        "name": "entity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateRootEntityRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Entity created successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/entity/sub": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new entity as a child of an existing entity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Entity Management"
                ],
                "summary": "Create a sub-entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cognito ID",
                        "name": "X-Cognito-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Entity information",
                        "name": "entity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSubEntityRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Sub-entity created successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Parent entity not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/has-entity": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Check if the authenticated user has any entities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Entity Management"
                ],
                "summary": "Check entity presence",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cognito ID",
                        "name": "X-Cognito-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entity presence check successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "boolean"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateRootEntityRequest": {
            "type": "object",
            "required": [
                "categoryId",
                "name"
            ],
            "properties": {
                "categoryId": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
        "dto.CreateSubEntityRequest": {
            "type": "object",
            "required": [
                "categoryId",
                "name"
            ],
            "properties": {
                "categoryId": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "parentEntityId": {
                    "description": "Optional. Only needed if adding an sub entity under another entity",
                    "type": "string"
                },
                "subUserID": {
                    "description": "UserID of the sub user. Optional. Only needed if adding a sub user",
                    "type": "string"
                }
            }
        },
...
        "dto.EntityChildrenResponse": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.EntityResponse"
                    }
                },
                "count": {
                    "type": "integer"
                },
                "parentId": {
                    "type": "string"
                }
            }
        },
        "dto.EntityHierarchyResponse": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "type": "string"
                },
                "categoryName": {
                    "type": "string"
                },
                "categoryType": {
                    "type": "string"
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.EntityHierarchyResponse"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "depth": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "dto.EntityResponse": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "type": "string"
                },
                "categoryName": {
                    "type": "string"
                },
                "categoryType": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "depth": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}